+++ modules/shiftmanager.js
